+++ docs/releases.md
@@ -18,6 +18,7 @@ changes over time.
 
 * github-226: Upgrade to Spark 3.2.2
 * github-227: [BUG] Flowman should not fail with field names containing "-", "/" etc
+* github-228: Padding and truncation of CHAR(n)/VARCHAR(n) should be configurable
 
 
 ### Version 0.26.0 - 2022-07-27